backend/requirements.txt
# FinBud Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
# Optional: Mentor Note (set OPENAI_API_KEY to enable)
openai>=1.12.0
# Dev / test
pytest>=7.4.0
